Replaced solids hierarchy on GPU with indexed one using LogicId/PlacedId.
This MR introduces a new hierarchical structure based on LogicId/PlacedId replacing the LogicalVolume/PlacedVolume one on the GPU. The new structures are index-based, and the IDs of placed and logical volumes are being preserved.

Navigation interfaces are changed by replacing `VPlacedVolume*` with `int`, so the id of the placed volume must be passed instead of the pointer.

The solid data model can be optionally NOT copied to the GPU, in Version 2 it will not be copied at all. If this is desired the sequence:
```
CudaManager::Instance().LoadGeometry(world);
CudaManager::Instance().Syncronize();
```
must be replaced by:
```
CudaManager::Instance().SynchronizeNavigationTable();
```
because the navigation table must be copied in all cases. If the solid data structures are not copied, currently all methods of the NavigationState object accessing solids data (Placed/LogicalVolumes) will fail.

New methods allowing index access in NavigationState:
```
static int WorldId(); // returns the placed volume id for the world volume
static int ToPlacedId(int placed); // returns the PlacedId struct corresponding to a placed volume index
void Push(int placed); // Pushes based on placed volume index
void PushDaughter(int id); // Pushes based of the daughter index of the current top volume
int TopId(); // returns the placed folume index of the top volume
int GetLastIdExited(); // returns the last exited placed volume id (see via SetLastExited)
```

testRaytracing now accepts a new parameter: only_surf, which activates skipping copying the solid model

Building with `VECGEOM_USE_NAVTUPLE=ON` is now the default mode

* [surface model] Replaced solids hierarchy on GPU with indexed one using LogicId/PlacedId.

Added solid types returned by unplaced volumes and stored in VolumeTree

Changed RelocatePointIn to use indices

LocatePointIn index based

Cleaner debug mode for LocatePointIn

Removed obsolete navigator

Fixes for the index-based navigation and testRaytacing able to run without copying the solid model to GPU.

NavStateIndex now also works without the solid model.

+++ VecGeom/surfaces/bvh/BVHsurf.h
@@ -446,8 +446,8 @@ public:
    * @returns Whether @p localpoint falls within a child element of this BVH.
    */
   template <typename Navigator>
-  VECCORE_ATT_HOST_DEVICE bool LevelLocate(long const exclude_item_id, Vector3D<Real_t> const &localpoint,
-                                           long &container_id, vecgeom::NavigationState &path) const
+  VECCORE_ATT_HOST_DEVICE bool LevelLocate(int const exclude_item_id, Vector3D<Real_t> const &localpoint,
+                                           int &container_id, vecgeom::NavigationState &path) const
   //  Vector3D<Real_t> &daughterlocalpoint) const
   {
     unsigned int stack[BVH_MAX_DEPTH], *ptr = &stack[1];
